+++ features/album/__init__.py
@@ -1,7 +1,15 @@
 """群相册功能模块。"""
 
 from .draw import MemeRenderer, detect_image_ext
+from .emoji_text import EmojiFont, TextPainter
 from .fonts import FontResolver
 from .service import AlbumFeature
 
-__all__ = ["AlbumFeature", "FontResolver", "MemeRenderer", "detect_image_ext"]
+__all__ = [
+    "AlbumFeature",
+    "EmojiFont",
+    "FontResolver",
+    "MemeRenderer",
+    "TextPainter",
+    "detect_image_ext",
+]